backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
python-whois==0.8.0
requests==2.32.3
urllib3==2.2.1
